refactor: move property hooks closer to implementation
