+++ packages/dataclass-editor/src/data/dataclass-utils.ts
@@ -1,4 +1,4 @@
-import { type DataClass, type DataClassType, type EntityClass } from './dataclass';
+import { type DataClass, type DataClassFieldModifier, type DataClassType, type EntityClass } from './dataclass';
 
 export const classTypeOf = (dataClass: DataClass): DataClassType => {
   if (dataClass.entity) {
@@ -12,4 +12,19 @@ export const classTypeOf = (dataClass: DataClass): DataClassType => {
 
 export const isEntity = (dataClass: DataClass): dataClass is EntityClass => {
   return !!dataClass.entity;
+};
+
+export const updateModifiers = (add: boolean, newModifiers: Array<DataClassFieldModifier>, modifier: DataClassFieldModifier) => {
+  if (add) {
+    if (modifier === 'ID' || modifier === 'VERSION') {
+      newModifiers = newModifiers.filter(mod => mod === 'GENERATED' || mod === 'PERSISTENT');
+    }
+    newModifiers.push(modifier);
+  } else {
+    if (modifier === 'ID') {
+      newModifiers = newModifiers.filter(mod => mod !== 'GENERATED');
+    }
+    newModifiers = newModifiers.filter(mod => mod !== modifier);
+  }
+  return newModifiers;
 };
