+++ packages/dataclass-editor/src/detail/field/entity/useFieldEntityProperty.test.ts
@@ -1,0 +1,18 @@
+import { customRenderHook } from '../../../context/test-utils/test-utils';
+import type { EntityClassField } from '../../../data/dataclass';
+import { useFieldEntityProperty } from './useFieldEntityProperty';
+
+test('useFieldEntityProperty', () => {
+  const field = { entity: { databaseName: 'databaseName' } } as EntityClassField;
+  let newField = {} as EntityClassField;
+  const view = customRenderHook(() => useFieldEntityProperty(), {
+    wrapperProps: { entityFieldContext: { field, setField: field => (newField = field) } }
+  });
+  expect(view.result.current.field).toEqual(field);
+
+  const originalField = structuredClone(field);
+  view.result.current.setProperty('databaseName', 'NewDatabaseName');
+  expect(field).toEqual(originalField);
+
+  expect(newField.entity.databaseName).toEqual('NewDatabaseName');
+});
